Switch from recipe microformat to microdata
The Big 3 (Google, Bing, Yahoo) are sponsoring a a change to microdata. Let's switch to that instead. All that's required is to add the appropriate data according to schema.org/Recipe, and remove the hrecipe class to avoid parser confusion. All other classes can remain for styling backcompat. Special attention will need to be paid to duration times, which will happen in future commit.